// angular-src/src/models/StatusList.ts
/*
 * @Description: Model for Status list
 * @Author: Guozhi Tang
 * @Date: 2019-07-24 10:50:13
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:45:10
 */
export class StatusList {
  moduleName: String;
  description: String;
  approber_id: Number;
  name: String;
  className: String;
  sop_id: Number;
  id: Number;
}