// angular-src/src/models/Operator.ts
/*
 * @Description: Model for Operator
 * @Author: Guozhi Tang
 * @Date: 2019-07-24 11:06:58
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:43:50
 */
export class Operator {
  moduleName: String;
  className: String;
  active: Boolean;
  admin: Boolean;
  id: Number;
  manufacturing: Boolean;
  name: String;
}