// angular-src/src/models/Department.ts
/*
 * @Description: Model for Department
 * @Author: Guozhi Tang
 * @Date: 2019-08-13 16:15:11
 * @Github: https://github.com/GuozhiTang/Bio-WebApp
 * @LastEditors: Guozhi Tang
 * @LastEditTime: 2019-08-30 14:43:12
 */
export class Department {
  id: Number;
  className: String;
  description: String;
  moduleName: String;
  name: String;
}